Change the -s option to "squeeze multiple blank lines" for pager
Change the -s option from "display charset Shift_JIS" to "squeeze multiple
blank lines" to work as /usr/bin/pager.  In addition, the options -j and -e
are disabled.  To specify the display charset, use -O{s|j|e} instead.

The patch was committed to the upstream on 2000-10-27.  However, it was
rejected because of Debian specific on 2010-08-03.