Fix table rowspan and colspan
rowspan and colspan should be at least 1. If either one of them is zero and the
other is larger than 1, HTT_X and HTT_Y attributes are not set correctly. Then
the calculation of maxcol or maxrow are wrong (not including colspan/rowspan).
w3m will overflow buffers later.

Examples to trigger this issue:
   echo '<table>0<td rowspan=0 colspan=300>a' | ./w3m -T text/html -dump
   echo '<table>0<td rowspan=0 colspan=30><img width=900000 src=0 height=0>'  | ./w3m -T text/html -dump