form.c: do not gratuitously turn GET into POST
When encountering a <form ...> tag that contains these values:

    method="get" enctype="multipart/form-data"

Do not transform the method into POST to accomodate enctype.

Instead behave in the compatible way that all other browsers
behave in this instance and ignore the enctype parameter
(treating it as the default application/x-www-form-urlencoded)
and perform a "GET" just as the method parameter requests.

This behavior produces far more compatible results than
gratuitously changing the "get" into a "post" which can
result in unexpected "405 Method Not Allowed" errors.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>