Add command to go home.
When w3m is launched, if no other options are specified, it attempts to
read HTTP_HOME and WWW_HOME from the environment and upon finding a
value for one of these load the url specified.  Once launched, though,
w3m provides no convenience for navigating to the home page.

Here, that ability is added.  A new command GOTO_HOME is defined with a
default key binding of C-_.