Reduce scope of dangerous macro
The Strcatc macro does no bounds checking when adding to the string.
This may lead to a out-of-bounds write violation.

It is only used in base64_encode where the needed size is known
beforehand. In this case it's ok (and cheaper) to omit the check.

Move the macro to the inner scope of the function to avoid programmers
using it.

While at it, do the same with the Strnulterm as it is only used once in
this function.